+++ src/components/NewsFeed/ShowPost.css
@@ -1,8 +1,13 @@
 #ShowPostContainer {
   width: 300px;
-  height: 300px;
+  height: 420px;
   border: 1px solid black;
   margin: 10px auto;
   padding: 10px;
   justify-content: center;
+}
+
+.photo {
+  height: 150px;
+  width: 300px;
 }
